added example2 and example2_group: 1 producer 2 consumers
added example2 and example2_group: 1 producer 2 consumers

Adds example with a single producer that appends records a sequence of numbers to a single topic evenly across all partitions.

Consumer consumes events from the topic and increments a count state variable.  After each batch is processed, the current count is spoked with espeak text-to-speech library.

Both example2 and example2_group have 2 consumers; but in one of them the consumers are in the same group set

note: for testing purposes consumers are allowed to create topics and the default parition count has been set to 2

removed maven dependency on espeak and added it to project because the reference implementation was not blocking the calling thread